Split URIResolver and DependencyLoader to separate classes
