+++ hummingbird-client/src/main/java/com/vaadin/client/URIResolver.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2000-2016 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.vaadin.client;
+
+import com.google.gwt.http.client.URL;
+import com.vaadin.shared.VaadinUriResolver;
+
+/**
+ * Client side URL resolver for vaadin protocols.
+ *
+ * @author Vaadin
+ * @since
+ */
+public class URIResolver extends VaadinUriResolver {
+    private transient Registry registry;
+
+    /**
+     * Creates a new instance connected to the given registry.
+     *
+     * @param registry
+     *            the global registry
+     */
+    public URIResolver(Registry registry) {
+        this.registry = registry;
+    }
+
+    @Override
+    protected String getVaadinDirUrl() {
+        return registry.getApplicationConfiguration().getVaadinDirUrl();
+    }
+
+    @Override
+    protected String getServiceUrl() {
+        return registry.getApplicationConfiguration().getServiceUrl();
+    }
+
+    @Override
+    protected String encodeQueryStringParameterValue(String queryString) {
+        return URL.encodeQueryString(queryString);
+    }
+
+}
